requirements.txt
pyyaml
ipydex
# owlready2
# rdflib
# pydantic
